portooBikee/porto-bike-care-react (2)/Sprint4 - IA/requirements.txt
opencv-python
werkzeug
numpy
flask
flask-cors
pillow
tensorflow